adding the components to the sidebar
