+++ frontend/src/App.jsx
@@ -28,6 +28,8 @@ import Med from "./pages/PharmacyPOV/med";
 import ViewMedReq from "./pages/PharmacyPOV/ViewMedReq";
 import ICUViewPatient from "./pages/ICU POV/ICUiewPatients";
 import OverallInventory from "./pages/Inventory/OverAllInventory";
+import PharmacyTransferHistory from "./pages/History/PharmacyTransferHistory";
+import CsrTransferHistory from "./pages/History/CsrTransferHistory";
 import { element } from "prop-types";
 
 function App() {
