+++ frontend/src/pages/Inventory/OverAllInventory.jsx
@@ -1,124 +1,26 @@
-import React, { useEffect, useState } from "react";
-import { ref, get, onValue } from "firebase/database";
-import { database } from "../../firebase/firebase";
+import { useState } from "react";
+import OverAllSupply from "./OverAllSupply";
+import OverAllMedicine from "./OverAllMedicine";
 
-const OverallInventory = () => {
-  const [overallInventory, setOverallInventory] = useState({});
-  const [searchTerm, setSearchTerm] = useState("");
+const OverAllInventory = () => {
 
-  useEffect(() => {
-    const suppliesRef = ref(database, "supplies");
-    const departmentsRef = ref(database, "departments");
-
-    const fetchSupplies = async () => {
-      try {
-        const suppliesSnapshot = await get(suppliesRef);
-        return suppliesSnapshot.exists() ? suppliesSnapshot.val() : {};
-      } catch (error) {
-        console.error("Error fetching supplies:", error);
-        return {};
-      }
-    };
-
-    const fetchDepartments = async () => {
-      try {
-        const departmentsSnapshot = await get(departmentsRef);
-        return departmentsSnapshot.exists() ? departmentsSnapshot.val() : {};
-      } catch (error) {
-        console.error("Error fetching departments:", error);
-        return {};
-      }
-    };
-
-    const fetchOverallInventory = async () => {
-      const [suppliesData, departmentsData] = await Promise.all([
-        fetchSupplies(),
-        fetchDepartments(),
-      ]);
-
-      const totalInventory = {};
-
-      // Process main supplies
-      Object.entries(suppliesData).forEach(([key, value]) => {
-        totalInventory[key] = {
-          itemName: value.itemName,
-          totalQuantity: value.quantity || 0, // Main node quantity
-        };
-      });
-
-      // Process each department's supplies under the 'localSupplies' node
-      Object.entries(departmentsData).forEach(
-        ([departmentKey, departmentValue]) => {
-          const departmentSupplies = departmentValue.localSupplies || {}; // Ensure you're targeting the correct supplies node within departments
-
-          Object.entries(departmentSupplies).forEach(([key, value]) => {
-            if (totalInventory[key]) {
-              // If the item exists in the main supplies, add the department's quantity
-              totalInventory[key].totalQuantity += value.quantity || 0;
-            } else {
-              // If the item is not in the main supplies, create a new entry
-              totalInventory[key] = {
-                itemName: value.itemName,
-                totalQuantity: value.quantity || 0, // Only department's quantity
-              };
-            }
-          });
-        }
-      );
-
-      setOverallInventory(totalInventory);
-    };
-
-    fetchOverallInventory();
-
-    const unsubscribeSupplies = onValue(suppliesRef, fetchOverallInventory);
-    const unsubscribeDepartments = onValue(
-      departmentsRef,
-      fetchOverallInventory
-    );
-
-    return () => {
-      unsubscribeSupplies();
-      unsubscribeDepartments();
-    };
-  }, []);
-
-  // Filter the inventory based on the search term
-  const filteredInventory = Object.entries(overallInventory).filter(
-    ([key, item]) =>
-      item.itemName.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
+  const [selectedTab, setSelectedTab] = useState("medicine")
   return (
-    <div className="max-w-full mx-auto mt-6 bg-white rounded-lg shadow-lg p-6">
-      <h1 className="text-xl font-bold mb-4">Overall Supplies Inventory</h1>
-      <input
-        type="text"
-        placeholder="Search for items..."
-        value={searchTerm}
-        onChange={(e) => setSearchTerm(e.target.value)}
-        className="border p-2 mb-4 w-full"
-      />
-      <table className="min-w-full border-collapse border border-gray-300">
-        <thead>
-          <tr>
-            <th className="border border-gray-300 p-2">Item Name</th>
-            <th className="border border-gray-300 p-2">Total Quantity</th>
-          </tr>
-        </thead>
-        <tbody>
-          {filteredInventory.map(([key, item]) => (
-            <tr key={key}>
-              <td className="border border-gray-300 p-2">{item.itemName}</td>
-              <td className="border border-gray-300 p-2">
-                {item.totalQuantity}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
+    <div>
+     <button onClick={() => setSelectedTab("medicine")} className={`space-x-4 px-6 py-2 rounded-md transition duration-200 ${
+              selectedTab === "medicine"
+                ? "bg-slate-900 text-white text-bold"
+                : "bg-slate-200 text-gray-900"
+            }`}>Medicine Overall Inventory</button>
+     <button onClick={() => setSelectedTab("supply")} className={`space-x-4 px-6 py-2 rounded-md transition duration-200 ${
+              selectedTab === "supply"
+                ? "bg-slate-900 text-white text-bold"
+                : "bg-slate-200 text-gray-900"
+            }`}>Supply Overall Inventory</button>
+     {selectedTab === "supply" && <OverAllSupply />}
+     {selectedTab === "medicine" && <OverAllMedicine />}
     </div>
   );
-};
+}
 
-export default OverallInventory;
+export default OverAllInventory;
